requirements.txt
selenium
pytest
webdriver-manager
allure-pytest
pytest-xdist
pytest-html
